Editing of my article
Use a dash to set off an abrupt break or interruption and to announce a long appositive or summary. line 4, 16
Use proper vocab: line 19 delete advance, add advantage line 38
Use the proper case of pronoun. line 4, 13, 41
The number of the subject determines the number of the verb. line 18
Place a comma before a conjunction introducing an independent clause. line 4, 5, 14, 41